requirements.txt
psycopg2-binary
python-dotenv
pandas
matplotlib
